+++ primefaces/src/main/java/org/primefaces/component/column/ColumnBase.java
@@ -56,6 +56,7 @@ public abstract class ColumnBase extends UIColumn implements org.primefaces.comp
         filterMatchMode,
         filterMaxLength,
         filterPosition,
+        filterPlaceholder,
         filterStyle,
         filterStyleClass,
         filterValue,
@@ -166,6 +167,15 @@ public abstract class ColumnBase extends UIColumn implements org.primefaces.comp
     @Override
     public String getFilterPosition() {
         return (String) getStateHelper().eval(PropertyKeys.filterPosition, "bottom");
+    }
+
+    @Override
+    public String getFilterPlaceholder() {
+        return (String) getStateHelper().eval(PropertyKeys.filterPlaceholder, null);
+    }
+
+    public void setFilterPlaceholder(String filterPlaceholder) {
+        getStateHelper().put(PropertyKeys.filterPlaceholder, filterPlaceholder);
     }
 
     public void setFilterPosition(String filterPosition) {