+++ primefaces/src/main/java/org/primefaces/component/datatable/DataTableRenderer.java
@@ -797,6 +797,10 @@ public class DataTableRenderer extends DataRenderer {
             writer.writeAttribute("maxlength", column.getFilterMaxLength(), null);
         }
 
+        if (column.getFilterPlaceholder() != null) {
+            writer.writeAttribute("placeholder", column.getFilterPlaceholder(), null);
+        }
+
         writer.endElement("input");
     }